requirements.txt
streamlit
numpy
Pillow
tensorflow==2.12.0
